+++ src/SFA.DAS.EmployerFinance.Jobs.UnitTests/SFA.DAS.EmployerFinance.Jobs.UnitTests.csproj
@@ -6,6 +6,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="AutoFixture" Version="4.11.0" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.3.0" />
     <PackageReference Include="Moq" Version="4.5.28" />
     <PackageReference Include="NUnit" Version="3.10.1" />
